% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{adaptNONMEMControlStream}
\alias{adaptNONMEMControlStream}
\title{Adapt NONMEM control stream by manipulating the source file.}
\usage{
adaptNONMEMControlStream(file, rem_rate, rem_abbr_replace)
}
\arguments{
\item{file}{control stream file name}

\item{rem_rate}{remove RATE in control stream automatically to avoid issues with Pharmpy.
Otherwise, it will look for the dataset and possibly adapt the ODE's to add the rates, default is FALSE}

\item{rem_abbr_replace}{remove section ABBREVIATED REPLACE, causing issue in import, default is FALSE}
}
\value{
nothing
}
\description{
Adapt NONMEM control stream by manipulating the source file.
}
