% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_conversion.R
\name{addconvertRecord}
\alias{addconvertRecord}
\title{Add record to the specified CAMPSIS model.}
\usage{
addconvertRecord(model, record, emptyRecord, parameters)
}
\arguments{
\item{model}{specified CAMPSIS model}

\item{record}{record to add}

\item{emptyRecord}{empty code record, already instantiated with the right type}

\item{parameters}{parameters}
}
\value{
updated CAMPSIS model
}
\description{
Add record to the specified CAMPSIS model.
}
