% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_conversion.R
\name{addMapping}
\alias{addMapping}
\title{Add mapping to parameters object.}
\usage{
addMapping(x, parameters, type)
}
\arguments{
\item{x}{mapping vector}

\item{parameters}{parameters object}

\item{type}{mapping vector type}
}
\value{
parameters
}
\description{
Add mapping to parameters object.
}
