% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmx_model_conversion.R
\name{addRecordToPmxModel}
\alias{addRecordToPmxModel}
\title{Add record to the specified PMX model.}
\usage{
addRecordToPmxModel(model, record, emptyRecord, parameters)
}
\arguments{
\item{model}{specified PMX model}

\item{record}{record to add}

\item{emptyRecord}{empty code record, already instantiated with the right type}

\item{parameters}{parameters}
}
\value{
updated PMX model
}
\description{
Add record to the specified PMX model.
}
