% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_extract_parameters.R
\name{autoExtractParameters2}
\alias{autoExtractParameters2}
\title{Auto-extra parameters from model code.
Unused model parameters are automatically deleted.
Indexes are automatically re-adjusted.}
\usage{
autoExtractParameters2(model)
}
\arguments{
\item{model}{any Campsis model. All parameters must be named (i.e. not NA)}
}
\value{
updated model
}
\description{
Auto-extra parameters from model code.
Unused model parameters are automatically deleted.
Indexes are automatically re-adjusted.
}
