% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_mapping.R
\name{autoRenameParameters}
\alias{autoRenameParameters}
\title{Auto-rename parameters based on equations.}
\usage{
autoRenameParameters(model, mapping)
}
\arguments{
\item{model}{CAMPSIS model}

\item{mapping}{mapping object}
}
\value{
CAMPSIS model with updated parameters
}
\description{
Auto-rename parameters based on equations.
}
