% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret_to_pow.R
\name{caretToPow}
\alias{caretToPow}
\title{Convert the caret operator into the pow function.}
\usage{
caretToPow(x)
}
\arguments{
\item{x}{object to transform}
}
\description{
Convert the caret operator into the pow function.
}
