% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret_to_pow.R
\name{caretToPowCore}
\alias{caretToPowCore}
\title{Convert the caret operator into the pow function.}
\usage{
caretToPowCore(x)
}
\arguments{
\item{x}{string to transform}
}
\value{
transformed expression
}
\description{
Convert the caret operator into the pow function.
}
