% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substitute_duplicates.R
\name{checkForDuplicateEquationNames}
\alias{checkForDuplicateEquationNames}
\title{Check for duplicate equation names.}
\usage{
checkForDuplicateEquationNames(model)
}
\arguments{
\item{model}{Campsis model}
}
\value{
Campsis model without duplicate equation names
}
\description{
Check for duplicate equation names.
}
