% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_conversion.R
\name{checkNewindStatement}
\alias{checkNewindStatement}
\title{Check if the equation is a newind statement.
E.g. if Equation("OCB", "first(CB, ID)"),
we return an if-statement.}
\usage{
checkNewindStatement(equation)
}
\arguments{
\item{equation}{equation to check}
}
\value{
IfStatement or Equation
}
\description{
Check if the equation is a newind statement.
E.g. if Equation("OCB", "first(CB, ID)"),
we return an if-statement.
}
