% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmx_model_conversion.R
\name{compartmentSystemToPmxModel}
\alias{compartmentSystemToPmxModel}
\title{Pharmpy compartment system conversion to PMX model.}
\usage{
compartmentSystemToPmxModel(system)
}
\arguments{
\item{system}{Pharmpy compartment system}
}
\value{
DES record (PMX domain)
}
\description{
Pharmpy compartment system conversion to PMX model.
}
