% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxode_conversion.R
\name{compartmentSystemToRxODE}
\alias{compartmentSystemToRxODE}
\title{Pharmpy compartment system conversion to RxODE code.}
\usage{
compartmentSystemToRxODE(system)
}
\arguments{
\item{system}{Pharmpy compartment system}
}
\value{
RxODE code
}
\description{
Pharmpy compartment system conversion to RxODE code.
}
