% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex_if_statement.R
\docType{class}
\name{complex_if_else_statement-class}
\alias{complex_if_else_statement-class}
\title{Complex if-else statement class.}
\description{
Complex if-else statement class.
}
\section{Slots}{

\describe{
\item{\code{list}}{list of IF statements}
}}

