% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex_if_statement.R
\name{ComplexIfElseStatement}
\alias{ComplexIfElseStatement}
\title{Create a new complex IF statement.}
\usage{
ComplexIfElseStatement()
}
\value{
an complex IF statement
}
\description{
Create a new complex IF statement.
}
