% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex_if_statement.R
\name{ComplexIfStatement}
\alias{ComplexIfStatement}
\title{Create a new complex IF statement.}
\usage{
ComplexIfStatement()
}
\value{
an complex IF statement
}
\description{
Create a new complex IF statement.
}
