% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_mapping.R
\name{containsParameter}
\alias{containsParameter}
\title{Contains parameter.}
\usage{
containsParameter(x, parameter)
}
\arguments{
\item{x}{any model statement}

\item{parameter}{parameter to search for a name in model}
}
\value{
a logical value
}
\description{
Contains parameter.
}
