% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_conversion.R
\name{convertCompartmentSystem}
\alias{convertCompartmentSystem}
\title{Pharmpy compartment system conversion to PMX model.}
\usage{
convertCompartmentSystem(model)
}
\arguments{
\item{model}{Pharmpy model}
}
\value{
ODE record (CAMPSIS domain)
}
\description{
Pharmpy compartment system conversion to PMX model.
}
