% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_from_rxode2.R
\name{convertComplexIfStatements}
\alias{convertComplexIfStatements}
\title{The goal is to convert a complex if statement structure into simple if statements.}
\usage{
convertComplexIfStatements(model)
}
\arguments{
\item{model}{original Campsis model}
}
\value{
a Campsis model without complex if statements
}
\description{
The goal is to convert a complex if statement structure into simple if statements.
}
