% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_conversion.R
\name{convertParameters}
\alias{convertParameters}
\title{Define and annotate your parameters.}
\usage{
convertParameters(model, mapping, estimate)
}
\arguments{
\item{model}{Pharmpy model}

\item{mapping}{PMX mapping}

\item{estimate}{if TRUE, estimated values are used, if FALSE, initial values are used}
}
\value{
parameters definition table
}
\description{
Define and annotate your parameters.
}
