% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_conversion.R
\name{convertPiecewise}
\alias{convertPiecewise}
\title{SymPy piecewise conversion to PMX model.}
\usage{
convertPiecewise(symbol, piecewise)
}
\arguments{
\item{symbol}{SymPy statement symbol}

\item{piecewise}{SymPy piecewise}
}
\value{
C code
}
\description{
SymPy piecewise conversion to PMX model.
}
