% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_conversion.R
\name{convertRecord}
\alias{convertRecord}
\title{NONMEM record (pharmpy) to CAMPSIS model.}
\usage{
convertRecord(record, emptyRecord, parameters)
}
\arguments{
\item{record}{one or more NONMEM record}

\item{emptyRecord}{empty code record, already instantiated with the right type}

\item{parameters}{parameters}
}
\value{
a CAMPSIS record
}
\description{
NONMEM record (pharmpy) to CAMPSIS model.
}
