% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varcov_conversion.R
\name{convertVarcov}
\alias{convertVarcov}
\title{Convert raw variance-covariance matrix for pmxmod.}
\usage{
convertVarcov(varcov, parameters)
}
\arguments{
\item{varcov}{raw variance-covariance matrix}

\item{parameters}{pmxmod parameters}
}
\value{
an updated variance-covariance matrix
}
\description{
Convert raw variance-covariance matrix for pmxmod.
}
