% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{covarToCor}
\alias{covarToCor}
\title{Convert off-diagonal elements to correlations.}
\usage{
covarToCor(model)
}
\arguments{
\item{model}{Campsis model}
}
\value{
updated Campsis model
}
\description{
Convert off-diagonal elements to correlations.
}
