% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmx_model_conversion.R
\name{desToPmxModel}
\alias{desToPmxModel}
\title{DES block (non processed) to PMX model.}
\usage{
desToPmxModel(desRecord)
}
\arguments{
\item{statements}{SymPy statements (not processed)}
}
\value{
C code
}
\description{
DES block (non processed) to PMX model.
}
