% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex_if_statement.R
\docType{class}
\name{else_if_statement-class}
\alias{else_if_statement-class}
\title{Else-if statement class.}
\description{
Else-if statement class.
}
\section{Slots}{

\describe{
\item{\code{list}}{list of IF statements}
}}

