% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex_if_statement.R
\name{ElseStatement}
\alias{ElseStatement}
\title{Create a new ELSE statement.}
\usage{
ElseStatement(statements)
}
\arguments{
\item{statements}{embedded statements}
}
\value{
an ELSE statement
}
\description{
Create a new ELSE statement.
}
