% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonmem_qualification_prep.R
\name{executeNONMEM}
\alias{executeNONMEM}
\title{Execute NONMEM. PsN will be called automatically by R.
Prepared control stream 'model.mod' is executed automatically and NONMEM results
are returned in the form of a data frame.}
\usage{
executeNONMEM(folder, reexecuteNONMEM = T, ctl_name = "model.mod")
}
\arguments{
\item{folder}{qualification folder where the control stream is}

\item{reexecuteNONMEM}{force re-execute NONMEM if results already exist}

\item{ctl_name}{control stream name, default is 'model.mod'}
}
\description{
Execute NONMEM. PsN will be called automatically by R.
Prepared control stream 'model.mod' is executed automatically and NONMEM results
are returned in the form of a data frame.
}
