% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonmem_qualification_prep.R
\name{executeSimulationCtl}
\alias{executeSimulationCtl}
\title{Prepare simulation control stream for qualification and execute it with NONMEM.}
\usage{
executeSimulationCtl(
  file,
  updateInits = TRUE,
  model,
  dataset,
  variables,
  folder,
  reexecuteNONMEM = TRUE
)
}
\arguments{
\item{file}{path to control stream file}

\item{updateInits}{update initial conditions in the control stream, default is TRUE (lst required)}

\item{model}{Campsis model that was previously imported (used for mapping the ETAs)}

\item{dataset}{simulation dataset, data frame}

\item{variables}{variables to output}

\item{folder}{where to write the control stream}

\item{reexecuteNONMEM}{re-execute NONMEM if results already exist, default is TRUE}
}
\value{
the path to the simulation control stream
}
\description{
Prepare simulation control stream for qualification and execute it with NONMEM.
}
