% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_conversion.R
\name{exportCampsisModel}
\alias{exportCampsisModel}
\title{Export CAMPSIS model.}
\usage{
exportCampsisModel(pharmpyModel, parameters, varcov, mapping)
}
\arguments{
\item{pharmpyModel}{Pharmpy model}

\item{parameters}{parameters (before auto mapping)}

\item{varcov}{varcov imported with Pharmpy}

\item{mapping}{initial mapping object}
}
\value{
the CAMPSIS model
}
\description{
Export CAMPSIS model.
}
