% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_extended_if_statement.R
\docType{class}
\name{extended_if_statement-class}
\alias{extended_if_statement-class}
\title{Extended if-statement class.}
\description{
Extended if-statement class.
}
\section{Slots}{

\describe{
\item{\code{condition}}{if-statement condition}

\item{\code{statements}}{list of statements}
}}

