% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_extended_if_statement.R
\name{ExtendedIfStatement}
\alias{ExtendedIfStatement}
\title{Create a new IF-statement.}
\usage{
ExtendedIfStatement(condition, statements)
}
\arguments{
\item{condition}{condition, single character string}

\item{statements}{list of statements}
}
\value{
an extended if-statement
}
\description{
Create a new IF-statement.
}
