% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_from_rxode2.R
\name{extractCompartmentPropertiesFromRxode}
\alias{extractCompartmentPropertiesFromRxode}
\title{Extract compartment properties from rxode2 statements.}
\usage{
extractCompartmentPropertiesFromRxode(model)
}
\arguments{
\item{model}{a Campsis model with all original rxode2 statements in the ODE block}
}
\value{
a Campsis model with updated compartment properties
}
\description{
Extract compartment properties from rxode2 statements.
}
