% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_from_rxode2.R
\name{extractInitialConditionsFromRxode}
\alias{extractInitialConditionsFromRxode}
\title{Extract initial conditions from rxode2 statements.}
\usage{
extractInitialConditionsFromRxode(model)
}
\arguments{
\item{model}{a Campsis model with all original rxode2 statements in the ODE block}
}
\value{
a Campsis model with updated initial conditions
}
\description{
Extract initial conditions from rxode2 statements.
}
