% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_from_rxode2.R
\name{extractModelCodeFromRxode}
\alias{extractModelCodeFromRxode}
\title{Extract model code from an rxode2 function object.
Note that all compartment names will be prefixed with 'A_'.}
\usage{
extractModelCodeFromRxode(rxmod)
}
\arguments{
\item{rxmod}{the rxode2 function object}
}
\value{
a Campsis model with all original rxode2 statements in the ODE block and detected compartments
}
\description{
Extract model code from an rxode2 function object.
Note that all compartment names will be prefixed with 'A_'.
}
