% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonmem_block.R
\name{extractNONMEMBlock}
\alias{extractNONMEMBlock}
\title{Extract NONMEM block.}
\usage{
extractNONMEMBlock(x, name, first = TRUE)
}
\arguments{
\item{x}{NONMEM control stream as a character string}

\item{name}{name of the NONMEM block to extract}

\item{first}{if TRUE, return only the first block found, otherwise return all blocks}
}
\value{
a list of nonmem_block objects, or a single nonmem_block object if first=TRUE
}
\description{
Extract NONMEM block.
}
