% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonmem_dataset_parser.R
\name{extractOptions}
\alias{extractOptions}
\title{Extract options from NONMEM block.}
\usage{
extractOptions(input)
}
\arguments{
\item{input}{input NONMEM block (i.e. INPUT, TABLE, etc.)}
}
\value{
a list of options (key/value form, value=NA if no value is provided)
}
\description{
Extract options from NONMEM block.
}
