% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_from_rxode2.R
\name{extractParametersFromRxode}
\alias{extractParametersFromRxode}
\title{Extract all parameters from an rxode2 function object.}
\usage{
extractParametersFromRxode(rxmod, pop_parameter_regex, omega_parameter_regex)
}
\arguments{
\item{rxmod}{the rxode2 function object}

\item{pop_parameter_regex}{regular expression to identify the population parameters}

\item{omega_parameter_regex}{regular expression to identify the OMEGA parameters}
}
\value{
a functional Campsis model
}
\description{
Extract all parameters from an rxode2 function object.
}
