% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharmpy_domain.R
\name{extractValueAfterUnderscore}
\alias{extractValueAfterUnderscore}
\title{Extract value(s) after underscore.}
\usage{
extractValueAfterUnderscore(str)
}
\arguments{
\item{str}{string}
}
\value{
a string
}
\description{
Extract value(s) after underscore.
}
