% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonmem_domain_utilities.R
\name{extractValueInParentheses}
\alias{extractValueInParentheses}
\title{Extract value between parentheses.}
\usage{
extractValueInParentheses(str)
}
\arguments{
\item{str}{string}
}
\value{
a string
}
\description{
Extract value between parentheses.
}
