% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonmem_dataset_parser.R
\name{getFirstRecord}
\alias{getFirstRecord}
\title{Get first NONMEM record from the NONMEM control stream for the given section name.}
\usage{
getFirstRecord(pharmpy, name, stop_if_not_found = TRUE)
}
\arguments{
\item{pharmpy}{pharmpy model}

\item{name}{NONMEM section name}

\item{stop_if_not_found}{throw an error if no section was found}
}
\value{
a record
}
\description{
Get first NONMEM record from the NONMEM control stream for the given section name.
}
