% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{getName}
\alias{getName}
\title{Get name in model code.}
\usage{
getName(object)
}
\arguments{
\item{object}{generic object}

\item{type}{parameter type}

\item{index}{parameter index}
}
\value{
name
}
\description{
Get name in model code.
}
