% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonmem_domain_utilities.R
\name{getNMParameterType}
\alias{getNMParameterType}
\title{Get type of NONMEM parameter.}
\usage{
getNMParameterType(str)
}
\arguments{
\item{str}{string}
}
\value{
the type of NONMEM parameter
}
\description{
Get type of NONMEM parameter.
}
