% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_library.R
\name{getNONMEMModelTemplate}
\alias{getNONMEMModelTemplate}
\title{Get the path to a NONMEM control stream template, embedded in this package.
This template file will be created in a temporary folder.
A CSV dataset will be created as well (needed for pharmpy).}
\usage{
getNONMEMModelTemplate(advan, trans)
}
\arguments{
\item{advan}{ADVAN number}

\item{trans}{TRANS number}
}
\value{
the path to this template
}
\description{
Get the path to a NONMEM control stream template, embedded in this package.
This template file will be created in a temporary folder.
A CSV dataset will be created as well (needed for pharmpy).
}
