% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonmem_domain.R
\name{getPharmpyParameterType}
\alias{getPharmpyParameterType}
\title{Get type of NONMEM parameter.}
\usage{
getPharmpyParameterType(str)
}
\arguments{
\item{str}{string}
}
\value{
the type of NONMEM parameter
}
\description{
Get type of NONMEM parameter.
}
