% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requirements.R
\name{getPharmpyRequirements}
\alias{getPharmpyRequirements}
\title{Get the Pharmpy requirements.}
\usage{
getPharmpyRequirements()
}
\value{
a character vector
}
\description{
Get the Pharmpy requirements.
}
