% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requirements.R
\name{getPharmpyRequirementsOld}
\alias{getPharmpyRequirementsOld}
\title{Get the Pharmpy requirements (old version).}
\usage{
getPharmpyRequirementsOld()
}
\value{
a character vector
}
\description{
Get the Pharmpy requirements (old version).
}
