% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requirements.R
\name{getPharmpyRequirementsUpdated}
\alias{getPharmpyRequirementsUpdated}
\title{Get the Pharmpy requirements (updated version).}
\usage{
getPharmpyRequirementsUpdated()
}
\value{
a character vector
}
\description{
Get the Pharmpy requirements (updated version).
}
