% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_configuration.R
\name{getPythonPath}
\alias{getPythonPath}
\title{Get path to python executable. This path can be changed in config.yml.}
\usage{
getPythonPath()
}
\value{
path to python
}
\description{
Get path to python executable. This path can be changed in config.yml.
}
