% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonmem_dataset_parser.R
\name{getRecordAt}
\alias{getRecordAt}
\title{Get indexed NONMEM record from the NONMEM control stream for the given section name.}
\usage{
getRecordAt(pharmpy, name, index = 1, stop_if_not_found = TRUE)
}
\arguments{
\item{pharmpy}{pharmpy model}

\item{name}{NONMEM section name}

\item{index}{index of the record, default is 1}

\item{stop_if_not_found}{throw an error if no section was found}
}
\value{
a record
}
\description{
Get indexed NONMEM record from the NONMEM control stream for the given section name.
}
