% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_from_nonmem.R
\name{heuristicMoveToError}
\alias{heuristicMoveToError}
\title{Heuristic move from the ODE record to the Error record.}
\usage{
heuristicMoveToError(model)
}
\arguments{
\item{model}{model with all statements in the ODE block}
}
\value{
updated model with an error block similar to NONMEM
}
\description{
Heuristic move from the ODE record to the Error record.
}
