% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_from_rxode2.R
\name{heuristicMoveToMain}
\alias{heuristicMoveToMain}
\title{Heuristic move from ODE to main.}
\usage{
heuristicMoveToMain(model)
}
\arguments{
\item{model}{model with all statements in the ODE block}
}
\value{
updated model with a MAIN block
}
\description{
Heuristic move from ODE to main.
}
