% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonmem_dataset_parser.R
\name{importDataset}
\alias{importDataset}
\title{Import a NONMEM dataset and prepare it for the qualification.}
\usage{
importDataset(campsistrans, covariates = NULL)
}
\arguments{
\item{campsistrans}{campsistrans object}

\item{covariates}{covariates vector. If provided, only these covariates are kept in dataset.
NULL is default (all covariates are kept)}
}
\value{
a data frame
}
\description{
Import a NONMEM dataset and prepare it for the qualification.
}
