% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonmem_dataset_parser.R
\name{importETAs}
\alias{importETAs}
\title{Import ETA's.}
\usage{
importETAs(x, file, model, id = "ID")
}
\arguments{
\item{x}{imported NONMEM dataset}

\item{file}{TAB file with estimated ETA's}

\item{model}{CAMPSIS model}

\item{id}{mapping ID name}
}
\value{
updated data frame
}
\description{
Import ETA's.
}
