% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_from_monolix.R
\name{importMonolix}
\alias{importMonolix}
\title{Extract the model from Monolix.}
\usage{
importMonolix(mlxtranFile, modelFile = NULL, parametersFile = NULL)
}
\arguments{
\item{modelFile}{path to model file, optional}

\item{parametersFile}{path to estimated parameters file, optional}

\item{path}{to mlxtran file}
}
\value{
a functional Campsis model
}
\description{
Extract the model from Monolix.
}
