% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmxtran.R
\name{importNONMEM}
\alias{importNONMEM}
\title{Import NONMEM control stream and results using Pharmpy.}
\usage{
importNONMEM(x, mapping = NULL, estimate = FALSE)
}
\arguments{
\item{x}{path to NONMEM control stream}

\item{mapping}{a possible PMX mapping object}

\item{estimate}{if TRUE, estimated values are imported, if FALSE, initial values are used}
}
\value{
a PMXtran object
}
\description{
Import NONMEM control stream and results using Pharmpy.
}
