% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_from_nonmem.R
\name{importNONMEM2}
\alias{importNONMEM2}
\title{Extract the model from NONMEM via nonmem2rx.}
\usage{
importNONMEM2(ctlFile, extFile = NULL, covFile = NULL)
}
\arguments{
\item{ctlFile}{path to control stream file}

\item{extFile}{path to parameter estimates file, optional}

\item{covFile}{path to variance-covariance matrix file, optional}
}
\value{
a functional Campsis model
}
\description{
Extract the model from NONMEM via nonmem2rx.
}
