% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonmem_predictions_parser.R
\name{importPredictions}
\alias{importPredictions}
\title{Import NONMEM predictions.}
\usage{
importPredictions(file, output, table_no = NULL, obs_only = TRUE)
}
\arguments{
\item{file}{path to NONMEM control stream (dataset path will be automatically extracted)}

\item{output}{output(s) of interest, character vector}

\item{table_no}{table number to look for the predictions, if NULL, first table
containing ID, TIME and the output(s) of interest will be used}

\item{obs_only}{if TRUE, only observations (MDV=0) will be returned, default is TRUE}
}
\value{
a data frame with ID and the output(s) of interest
}
\description{
Import NONMEM predictions.
}
