% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_from_rxode2.R
\name{importRxode2}
\alias{importRxode2}
\title{Extract model code from an rxode2 function object.}
\usage{
importRxode2(
  rxmod,
  pop_parameter_regex = NULL,
  omega_parameter_regex = NULL,
  subroutine = NULL
)
}
\arguments{
\item{rxmod}{the rxode2 function object}

\item{pop_parameter_regex}{regular expression to identify the population parameters}

\item{omega_parameter_regex}{regular expression to identify the OMEGA parameters}

\item{subroutine}{optional subroutine, integer vector of 2 values: ADVAN and TRANS.}
}
\value{
a functional Campsis model
}
\description{
Extract model code from an rxode2 function object.
}
