% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharmpy_config.R
\name{installPharmpy}
\alias{installPharmpy}
\title{Install Pharmpy environment from R.}
\usage{
installPharmpy(config = UpdatedPharmpyConfig())
}
\arguments{
\item{config}{Pharmpy configuration object, default is UpdatedPharmpyConfig()}
}
\value{
virtual environment
}
\description{
Install Pharmpy environment from R.
}
