% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharmpy_config.R
\name{installPharmpyInBackground}
\alias{installPharmpyInBackground}
\title{Install python in the background.}
\usage{
installPharmpyInBackground(config = UpdatedPharmpyConfig())
}
\arguments{
\item{config}{Pharmpy configuration object, default is UpdatedPharmpyConfig()}
}
\value{
callr S4 object
}
\description{
Install python in the background.
}
