% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_configuration.R
\name{installPython}
\alias{installPython}
\title{Install python environment and pharmpy.}
\usage{
installPython(envname, python)
}
\arguments{
\item{envname}{name of the virtual environment name}

\item{python}{path to distribution of python (3.9 is recommended)}
}
\value{
virtual environment
}
\description{
Install python environment and pharmpy.
}
