% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_configuration.R
\name{installPythonInBackground}
\alias{installPythonInBackground}
\title{Install python in the background.}
\usage{
installPythonInBackground()
}
\value{
callr S4 object
}
\description{
Install python in the background.
}
