% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonmem_domain.R
\name{isNMArrayParameter}
\alias{isNMArrayParameter}
\title{Is NONMEM/Fortran single array parameter.}
\usage{
isNMArrayParameter(str, type)
}
\arguments{
\item{str}{string}

\item{type}{parameter type}
}
\value{
a logical value
}
\description{
Is NONMEM/Fortran single array parameter.
}
