% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonmem_domain.R
\name{isNMCompartmentVariable}
\alias{isNMCompartmentVariable}
\title{Is NONMEM compartment variable (A Fortran array).}
\usage{
isNMCompartmentVariable(str)
}
\arguments{
\item{str}{string}
}
\value{
a logical value
}
\description{
Is NONMEM compartment variable (A Fortran array).
}
