% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonmem_domain_utilities.R
\name{isNMDoubleArrayParameter}
\alias{isNMDoubleArrayParameter}
\title{Is NONMEM/Fortran double array parameter.}
\usage{
isNMDoubleArrayParameter(str, type)
}
\arguments{
\item{str}{string}

\item{type}{parameter type}
}
\value{
a logical value
}
\description{
Is NONMEM/Fortran double array parameter.
}
