% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonmem_domain_utilities.R
\name{isNMErrorParameter}
\alias{isNMErrorParameter}
\title{Is NONMEM ERROR parameter method (EPS or ERR Fortran array).}
\usage{
isNMErrorParameter(str)
}
\arguments{
\item{str}{string}
}
\value{
a logical value
}
\description{
Is NONMEM ERROR parameter method (EPS or ERR Fortran array).
}
