% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonmem_domain_utilities.R
\name{isNMEtaParameter}
\alias{isNMEtaParameter}
\title{Is NONMEM ETA parameter method (ETA Fortran array).}
\usage{
isNMEtaParameter(str)
}
\arguments{
\item{str}{string}
}
\value{
a logical value
}
\description{
Is NONMEM ETA parameter method (ETA Fortran array).
}
