% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonmem_domain.R
\name{isNMOmegaParameter}
\alias{isNMOmegaParameter}
\title{Is NONMEM OMEGA parameter method (OMEGA Fortran double array).}
\usage{
isNMOmegaParameter(str)
}
\arguments{
\item{str}{string}
}
\value{
a logical value
}
\description{
Is NONMEM OMEGA parameter method (OMEGA Fortran double array).
}
