% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonmem_domain.R
\name{isNMParameter}
\alias{isNMParameter}
\title{Is NONMEM parameter method.}
\usage{
isNMParameter(str)
}
\arguments{
\item{str}{string}
}
\value{
a logical value
}
\description{
Is NONMEM parameter method.
}
