% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonmem_domain.R
\name{isNMSigmaParameter}
\alias{isNMSigmaParameter}
\title{Is NONMEM SIGMA parameter method (SIGMA Fortran double array).}
\usage{
isNMSigmaParameter(str)
}
\arguments{
\item{str}{string}
}
\value{
a logical value
}
\description{
Is NONMEM SIGMA parameter method (SIGMA Fortran double array).
}
