% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonmem_domain.R
\name{isNMThetaParameter}
\alias{isNMThetaParameter}
\title{Is NONMEM THETA parameter method (THETA Fortran array).}
\usage{
isNMThetaParameter(str)
}
\arguments{
\item{str}{string}
}
\value{
a logical value
}
\description{
Is NONMEM THETA parameter method (THETA Fortran array).
}
