% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharmpy_domain.R
\name{isPharmpyArrayParameter}
\alias{isPharmpyArrayParameter}
\title{Is NONMEM/Fortran single array parameter.}
\usage{
isPharmpyArrayParameter(str, type)
}
\arguments{
\item{str}{string}

\item{type}{parameter type}
}
\value{
a logical value
}
\description{
Is NONMEM/Fortran single array parameter.
}
