% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharmpy_domain.R
\name{isPharmpyEpsParameter}
\alias{isPharmpyEpsParameter}
\title{Is NONMEM ERROR parameter method (EPS or ERR Fortran array).}
\usage{
isPharmpyEpsParameter(str)
}
\arguments{
\item{str}{string}
}
\value{
a logical value
}
\description{
Is NONMEM ERROR parameter method (EPS or ERR Fortran array).
}
