% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharmpy_domain.R
\name{isPharmpyEtaParameter}
\alias{isPharmpyEtaParameter}
\title{Is NONMEM ETA parameter method (ETA Fortran array).}
\usage{
isPharmpyEtaParameter(str)
}
\arguments{
\item{str}{string}
}
\value{
a logical value
}
\description{
Is NONMEM ETA parameter method (ETA Fortran array).
}
