% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharmpy_domain.R
\name{isPharmpyOmegaParameter}
\alias{isPharmpyOmegaParameter}
\title{Is NONMEM OMEGA parameter method (OMEGA Fortran double array).}
\usage{
isPharmpyOmegaParameter(str)
}
\arguments{
\item{str}{string}
}
\value{
a logical value
}
\description{
Is NONMEM OMEGA parameter method (OMEGA Fortran double array).
}
