% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharmpy_domain.R
\name{isPharmpyParameter}
\alias{isPharmpyParameter}
\title{Is NONMEM parameter method.}
\usage{
isPharmpyParameter(str)
}
\arguments{
\item{str}{string}
}
\value{
a logical value
}
\description{
Is NONMEM parameter method.
}
