% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharmpy_domain.R
\name{isPharmpySigmaParameter}
\alias{isPharmpySigmaParameter}
\title{Is NONMEM SIGMA parameter method (SIGMA Fortran double array).}
\usage{
isPharmpySigmaParameter(str)
}
\arguments{
\item{str}{string}
}
\value{
a logical value
}
\description{
Is NONMEM SIGMA parameter method (SIGMA Fortran double array).
}
