% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharmpy_domain.R
\name{isPharmpyThetaParameter}
\alias{isPharmpyThetaParameter}
\title{Is NONMEM THETA parameter method (THETA Fortran array).}
\usage{
isPharmpyThetaParameter(str)
}
\arguments{
\item{str}{string}
}
\value{
a logical value
}
\description{
Is NONMEM THETA parameter method (THETA Fortran array).
}
