% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonmem_qualification_prep.R
\name{loadCtl}
\alias{loadCtl}
\title{Load control stream with Pharmpy.}
\usage{
loadCtl(path, estimate)
}
\arguments{
\item{path}{path to original control stream}

\item{estimate}{reuse estimated parameters from the model fit, default is TRUE}
}
\value{
the updated campsistrans object
}
\description{
Load control stream with Pharmpy.
}
