% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{mapping}
\alias{mapping}
\title{Create PMX mapping.}
\usage{
mapping(theta = NULL, omega = NULL, sigma = NULL)
}
\arguments{
\item{theta}{named integer vector for THETA mapping}

\item{omega}{named integer vector for THETA mapping}

\item{sigma}{named integer vector for OMEGA mapping}
}
\value{
PMX mapping object
}
\description{
Create PMX mapping.
}
