% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmxtran.R
\name{modelToNONMEM}
\alias{modelToNONMEM}
\title{Write PMX model to NONMEM control stream file.}
\usage{
modelToNONMEM(model, path)
}
\arguments{
\item{model}{PMX model}

\item{path}{path to desired control stream file}
}
\description{
Write PMX model to NONMEM control stream file.
}
