% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_conversion.R
\name{moveInitialConditions}
\alias{moveInitialConditions}
\title{Move initial conditions from MAIN to INIT section.}
\usage{
moveInitialConditions(model)
}
\arguments{
\item{model}{CAMPSIS model}
}
\value{
updated CAMPSIS model
}
\description{
Move initial conditions from MAIN to INIT section.
}
