% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{nameCovariance}
\alias{nameCovariance}
\title{Give name to the off-diagonal elements of the covariance matrix.}
\usage{
nameCovariance(model)
}
\arguments{
\item{model}{Campsis model}
}
\value{
updated Campsis model
}
\description{
Give name to the off-diagonal elements of the covariance matrix.
}
