% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_mapping.R
\name{needsAutoRenaming}
\alias{needsAutoRenaming}
\title{Check if the given parameter needs to be renamed.}
\usage{
needsAutoRenaming(parameter)
}
\arguments{
\item{parameter}{a parameter}
}
\value{
a logical value
}
\description{
Check if the given parameter needs to be renamed.
}
