% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonmem_update.R
\name{omegaSigmaToZero}
\alias{omegaSigmaToZero}
\title{Set OMEGA and SIGMA initial values to 0 and fix them.}
\usage{
omegaSigmaToZero(pmxtran)
}
\arguments{
\item{pmxtran}{PMXtran object}
}
\value{
the updated PMX model
}
\description{
Set OMEGA and SIGMA initial values to 0 and fix them.
}
