% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_definition.R
\name{params}
\alias{params}
\title{Define and annotate your parameters.}
\usage{
params(model, mapping = NULL)
}
\arguments{
\item{model}{Pharmpy model}

\item{mapping}{PMX mapping}
}
\value{
parameters definition table
}
\description{
Define and annotate your parameters.
}
