% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmx_model_conversion.R
\name{piecewiseToPmxModel}
\alias{piecewiseToPmxModel}
\title{SymPy piecewise conversion to PMX model.}
\usage{
piecewiseToPmxModel(symbol, piecewise)
}
\arguments{
\item{symbol}{SymPy statement symbol}

\item{piecewise}{SymPy piecewise}
}
\value{
C code
}
\description{
SymPy piecewise conversion to PMX model.
}
