% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxode_conversion.R
\name{piecewiseToRxODE}
\alias{piecewiseToRxODE}
\title{SymPy piecewise conversion to RxODE code.}
\usage{
piecewiseToRxODE(symbol, piecewise)
}
\arguments{
\item{symbol}{SymPy statement symbol}

\item{piecewise}{SymPy piecewise}
}
\value{
RxODE code
}
\description{
SymPy piecewise conversion to RxODE code.
}
