% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonmem_qualification_prep.R
\name{prepareNONMEMFiles}
\alias{prepareNONMEMFiles}
\title{Export campsistrans object to NONMEM control stream for qualification.
If NONMEM results are provided, they will replace the previous original values in the control stream.
OMEGA and SIGMA initial values will then be fixed and set to 0.
ETA arrays in control stream will be replaced by ETA covariates (e.g. ETA(1) -> ETA_1)}
\usage{
prepareNONMEMFiles(x, dataset, variables, compartments = NULL, outputFolder)
}
\arguments{
\item{x}{campsistrans object}

\item{dataset}{dataframe or tibble}

\item{variables}{variables to output (note: ID, ARM, TIME, EVID, MDV, DV, AMT, CMT, DOSENO are output by default)}

\item{compartments}{compartment indexes to output, numeric vector}

\item{outputFolder}{output folder to export the qualification control stream and CSV dataset}
}
\value{
the updated campsistrans object
}
\description{
Export campsistrans object to NONMEM control stream for qualification.
If NONMEM results are provided, they will replace the previous original values in the control stream.
OMEGA and SIGMA initial values will then be fixed and set to 0.
ETA arrays in control stream will be replaced by ETA covariates (e.g. ETA(1) -> ETA_1)
}
