% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{printSymPy}
\alias{printSymPy}
\title{Print symbolic expressions or conditions.}
\usage{
printSymPy(x, output = "C")
}
\arguments{
\item{x}{SymPy expression or condition}

\item{output}{type of desired output}
}
\value{
code
}
\description{
Print symbolic expressions or conditions.
}
