% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_conversion.R
\name{processParameters}
\alias{processParameters}
\title{Process parameters before return.}
\usage{
processParameters(parameters)
}
\arguments{
\item{parameters}{parameters}
}
\value{
updated parameters
}
\description{
Process parameters before return.
}
