% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonmem_qualification.R
\name{qualify}
\alias{qualify}
\title{Export PMXtran object to NONMEM control stream for qualification.
If NONMEM results are provided, they will replace the previous original values in the control stream.
OMEGA and SIGMA initial values will then be fixed and set to 0.
ETA arrays in control stream will be replaced by ETA covariates (e.g. ETA(1) -> ETA_1) if required.}
\usage{
qualify(pmxtran, etasAsCovariates = TRUE)
}
\arguments{
\item{pmxtran}{PMXtran object}

\item{etasAsCovariates}{ETA's as covariates, logical value}
}
\value{
the updated PMXtran object
}
\description{
Export PMXtran object to NONMEM control stream for qualification.
If NONMEM results are provided, they will replace the previous original values in the control stream.
OMEGA and SIGMA initial values will then be fixed and set to 0.
ETA arrays in control stream will be replaced by ETA covariates (e.g. ETA(1) -> ETA_1) if required.
}
