% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonmem_table_parser.R
\name{read.nonmemtable}
\alias{read.nonmemtable}
\title{Read a NONMEM table.}
\usage{
read.nonmemtable(content, varcov = FALSE)
}
\arguments{
\item{content}{table content}

\item{varvov}{must be set to TRUE for .cov files, default is FALSE. It simply indicates
the first column is not numeric.}
}
\value{
a dataframe
}
\description{
Read a NONMEM table.
}
