% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonmem_results_parser.R
\name{read.varcov}
\alias{read.varcov}
\title{Read NONMEM variance-covariance matrix files (extension .cov).}
\usage{
read.varcov(file)
}
\arguments{
\item{file}{the cov file to be read}
}
\value{
the raw variance covariance matrix, matrix
}
\description{
Read NONMEM variance-covariance matrix files (extension .cov).
}
