% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_conversion.R
\name{recordToPmxModel}
\alias{recordToPmxModel}
\title{NONMEM record (pharmpy) to PMX model.}
\usage{
recordToPmxModel(records, emptyRecord, parameters)
}
\arguments{
\item{records}{one or more NONMEM record}

\item{emptyRecord}{empty code record, already instantiated with the right type}

\item{parameters}{parameters}
}
\value{
a PMX record
}
\description{
NONMEM record (pharmpy) to PMX model.
}
