% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{removeAbbreviatedReplaceFromString}
\alias{removeAbbreviatedReplaceFromString}
\title{Remove $ABBREVIATED REPLACE section from string.}
\usage{
removeAbbreviatedReplaceFromString(x)
}
\arguments{
\item{x}{string value (the whole control stream, 1 line)}
}
\value{
the same string value without the abbreviated replace section
}
\description{
Remove $ABBREVIATED REPLACE section from string.
}
