% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonmem_block.R
\name{removeNONMEMBlock}
\alias{removeNONMEMBlock}
\title{Remove NONMEM block.}
\usage{
removeNONMEMBlock(x, name, first = TRUE)
}
\arguments{
\item{x}{NONMEM control stream as a character string}

\item{name}{name of the NONMEM block to extract}

\item{first}{if TRUE, only the first block will be removed, otherwise all blocks will be removed}
}
\value{
control stream as a character string
}
\description{
Remove NONMEM block.
}
