% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{removeNONMEMComments}
\alias{removeNONMEMComments}
\title{Remove NONMEM comments.}
\usage{
removeNONMEMComments(lines)
}
\arguments{
\item{lines}{lines}
}
\value{
lines without any comment
}
\description{
Remove NONMEM comments.
}
