% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_conversion.R
\name{removePiecewiseStatements}
\alias{removePiecewiseStatements}
\title{Remove the piecewise statements added by Pharmpy.}
\usage{
removePiecewiseStatements(model)
}
\arguments{
\item{model}{CAMPSIS model}
}
\value{
updated CAMPSIS model
}
\description{
Remove the piecewise statements added by Pharmpy.
}
