% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{removeRateFromCtl}
\alias{removeRateFromCtl}
\title{Remove RATE input from $INPUT field in given control stream.}
\usage{
removeRateFromCtl(file)
}
\arguments{
\item{file}{control stream file name}
}
\value{
nothing
}
\description{
Remove RATE input from $INPUT field in given control stream.
}
