% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{removeRateFromString}
\alias{removeRateFromString}
\title{Remove RATE input from $INPUT field (string-based model).}
\usage{
removeRateFromString(x)
}
\arguments{
\item{x}{string value (the whole control stream)}
}
\value{
the same string value without RATE input
}
\description{
Remove RATE input from $INPUT field (string-based model).
}
