% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_conversion.R
\name{removeUselessEquations}
\alias{removeUselessEquations}
\title{Remove useless equations (e.g. ETA_CL=ETA_CL). This can be useful when
the auto-mapping is used.}
\usage{
removeUselessEquations(model)
}
\arguments{
\item{model}{CAMPSIS model}
}
\value{
updated CAMPSIS model
}
\description{
Remove useless equations (e.g. ETA_CL=ETA_CL). This can be useful when
the auto-mapping is used.
}
