% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonmem_block.R
\name{replaceNONMEMBlock}
\alias{replaceNONMEMBlock}
\title{Replace NONMEM block.}
\usage{
replaceNONMEMBlock(x, name, content)
}
\arguments{
\item{x}{NONMEM control stream as a character string}

\item{name}{name of the NONMEM block to replace}

\item{content}{updated content, character vector}
}
\value{
control stream as a character string
}
\description{
Replace NONMEM block.
}
