% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbol_conversion.R
\name{replaceSymbol}
\alias{replaceSymbol}
\title{Replace all occurrences of the given symbol by the replacement symbol.}
\usage{
replaceSymbol(expression, symbol, replacementSymbol)
}
\arguments{
\item{expression}{SymPy expression}

\item{symbol}{SymPy symbol to be replaced}

\item{replacementSymbol}{SymPy replacement symbol}
}
\value{
the updated expression
}
\description{
Replace all occurrences of the given symbol by the replacement symbol.
}
