% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmxtran.R
\name{replaceSymbolAuto}
\alias{replaceSymbolAuto}
\title{Replace all occurrences of the given symbol according to the parameters definition table.}
\usage{
replaceSymbolAuto(expression, symbol, params)
}
\arguments{
\item{expression}{SymPy expression}

\item{symbol}{SymPy symbol to be replaced}

\item{params}{parameters definition table}
}
\value{
the updated expression
}
\description{
Replace all occurrences of the given symbol according to the parameters definition table.
}
