% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbol_conversion.R
\name{retrieveCompartmentName}
\alias{retrieveCompartmentName}
\title{Retrieve compartment name based on left hand side expression.}
\usage{
retrieveCompartmentName(lhs)
}
\arguments{
\item{lhs}{left hand side expression of the derivative}
}
\value{
the compartment name
}
\description{
Retrieve compartment name based on left hand side expression.
}
