% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_conversion.R
\name{retrieveInitialValues}
\alias{retrieveInitialValues}
\title{Retrieve initial values from Pharmpy parameter set.}
\usage{
retrieveInitialValues(model)
}
\arguments{
\item{model}{Pharmpy model}
}
\value{
S4 parameters object
}
\description{
Retrieve initial values from Pharmpy parameter set.
}
