% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_model_rxode2.R
\docType{class}
\name{rxode2_error_model-class}
\alias{rxode2_error_model-class}
\title{Rxode2 error model class.}
\description{
Rxode2 error model class.
}
\section{Slots}{

\describe{
\item{\code{add}}{additive error related variable}

\item{\code{prop}}{proportional error related variable}

\item{\code{combined1}}{combined1 enabled}

\item{\code{combined2}}{combined2 enabled}

\item{\code{endpoint}}{related endpoint}
}}

