% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_model_rxode2_parser.R
\name{Rxode2ErrorModelLexer}
\alias{Rxode2ErrorModelLexer}
\title{Rxode2 error model lexer.}
\value{
a lexer object
}
\description{
Rxode2ErrorModelLexer class to handle the lexer for Rxode2 error model
}
\section{Public fields}{
  \if{html}{\out{<div class="r6-fields">}}
  \describe{
    \item{\code{tokens}}{list of tokens}

    \item{\code{debug}}{logical value to enable/disable debug messages}

    \item{\code{t_ignore}}{function to ignore whitespace}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-Lexer-t_ENDPOINT}{\code{Lexer$t_ENDPOINT()}}
    \item \href{#method-Lexer-t_ADD}{\code{Lexer$t_ADD()}}
    \item \href{#method-Lexer-t_PROP}{\code{Lexer$t_PROP()}}
    \item \href{#method-Lexer-t_COMBINED1}{\code{Lexer$t_COMBINED1()}}
    \item \href{#method-Lexer-t_COMBINED2}{\code{Lexer$t_COMBINED2()}}
    \item \href{#method-Lexer-t_newline}{\code{Lexer$t_newline()}}
    \item \href{#method-Lexer-t_error}{\code{Lexer$t_error()}}
    \item \href{#method-Lexer-clone}{\code{Lexer$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lexer-t_ENDPOINT"></a>}}
\if{latex}{\out{\hypertarget{method-Lexer-t_ENDPOINT}{}}}
\subsection{\code{Lexer$t_ENDPOINT()}}{
  tokens list of tokens


  debug logical value to enable/disable debug messages


  t_ENDPOINT function to handle ENDPOINT token
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Lexer$t_ENDPOINT(re = "[a-zA-Z_][a-zA-Z0-9_\\\\.]*\\\\s*~\\\\s*", t)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{re}}{regular expression to match ENDPOINT token}
      \item{\code{t}}{token object}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lexer-t_ADD"></a>}}
\if{latex}{\out{\hypertarget{method-Lexer-t_ADD}{}}}
\subsection{\code{Lexer$t_ADD()}}{
  t_ADD function to handle ADD token
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Lexer$t_ADD(re = "add\\\\s*\\\\([^\\\\)]*\\\\)", t)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{re}}{regular expression to match ADD token}
      \item{\code{t}}{token object}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lexer-t_PROP"></a>}}
\if{latex}{\out{\hypertarget{method-Lexer-t_PROP}{}}}
\subsection{\code{Lexer$t_PROP()}}{
  t_PROP function to handle PROP token
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Lexer$t_PROP(re = "prop\\\\s*\\\\([^\\\\)]*\\\\)", t)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{re}}{regular expression to match PROP token}
      \item{\code{t}}{token object}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lexer-t_COMBINED1"></a>}}
\if{latex}{\out{\hypertarget{method-Lexer-t_COMBINED1}{}}}
\subsection{\code{Lexer$t_COMBINED1()}}{
  t_COMBINED1 function to handle COMBINED1 token
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Lexer$t_COMBINED1(re = "combined1\\\\s*\\\\(\\\\s*\\\\)", t)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{re}}{regular expression to match COMBINED1 token}
      \item{\code{t}}{token object}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lexer-t_COMBINED2"></a>}}
\if{latex}{\out{\hypertarget{method-Lexer-t_COMBINED2}{}}}
\subsection{\code{Lexer$t_COMBINED2()}}{
  t_COMBINED2 function to handle COMBINED2 token
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Lexer$t_COMBINED2(re = "combined2\\\\s*\\\\(\\\\s*\\\\)", t)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{re}}{regular expression to match COMBINED2 token}
      \item{\code{t}}{token object}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lexer-t_newline"></a>}}
\if{latex}{\out{\hypertarget{method-Lexer-t_newline}{}}}
\subsection{\code{Lexer$t_newline()}}{
  t_ignore function to ignore whitespace


  t_newline function to handle newlines
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Lexer$t_newline(re = "\\\\n+", t)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{re}}{regular expression to match newlines}
      \item{\code{t}}{token object}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lexer-t_error"></a>}}
\if{latex}{\out{\hypertarget{method-Lexer-t_error}{}}}
\subsection{\code{Lexer$t_error()}}{
  t_error function to handle errors
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Lexer$t_error(t)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{t}}{token object}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lexer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Lexer-clone}{}}}
\subsection{\code{Lexer$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Lexer$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
