% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxode2_parser.R
\name{Rxode2Parser}
\alias{Rxode2Parser}
\title{Rxode2 parser.}
\value{
a parser object
}
\description{
Rxode2Parser class to handle the parser for RxODE2
}
\section{Public fields}{
  \if{html}{\out{<div class="r6-fields">}}
  \describe{
    \item{\code{tokens}}{list of tokens}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-Parser-p_model_statements}{\code{Parser$p_model_statements()}}
    \item \href{#method-Parser-p_model_statement}{\code{Parser$p_model_statement()}}
    \item \href{#method-Parser-p_complex_if_else_statement}{\code{Parser$p_complex_if_else_statement()}}
    \item \href{#method-Parser-p_simple_if_statement}{\code{Parser$p_simple_if_statement()}}
    \item \href{#method-Parser-p_if_statement}{\code{Parser$p_if_statement()}}
    \item \href{#method-Parser-p_equation}{\code{Parser$p_equation()}}
    \item \href{#method-Parser-p_ode}{\code{Parser$p_ode()}}
    \item \href{#method-Parser-p_line_break}{\code{Parser$p_line_break()}}
    \item \href{#method-Parser-p_comment}{\code{Parser$p_comment()}}
    \item \href{#method-Parser-p_unknown_statement}{\code{Parser$p_unknown_statement()}}
    \item \href{#method-Parser-p_error}{\code{Parser$p_error()}}
    \item \href{#method-Parser-clone}{\code{Parser$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parser-p_model_statements"></a>}}
\if{latex}{\out{\hypertarget{method-Parser-p_model_statements}{}}}
\subsection{\code{Parser$p_model_statements()}}{
  tokens list of tokens


  function to handle model_statements
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Parser$p_model_statements(
  doc =
    "model_statements : model_statement\\n                         | model_statements model_statement",
  p
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{doc}}{doc argument}
      \item{\code{p}}{parser object}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parser-p_model_statement"></a>}}
\if{latex}{\out{\hypertarget{method-Parser-p_model_statement}{}}}
\subsection{\code{Parser$p_model_statement()}}{
  function to handle model_statement
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Parser$p_model_statement(
  doc =
    "model_statement : complex_if_else_statement\\n                         | equation \\n                         | ode\\n                         | line_break\\n                         | comment\\n                         | simple_if_statement\\n                         | unknown_statement",
  p
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{doc}}{doc argument}
      \item{\code{p}}{parser object}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parser-p_complex_if_else_statement"></a>}}
\if{latex}{\out{\hypertarget{method-Parser-p_complex_if_else_statement}{}}}
\subsection{\code{Parser$p_complex_if_else_statement()}}{
  function to handle complex_if_else_statement
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Parser$p_complex_if_else_statement(
  doc =
    "complex_if_else_statement : if_statement\\n                         | complex_if_else_statement if_statement",
  p
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{doc}}{doc argument}
      \item{\code{p}}{parser object}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parser-p_simple_if_statement"></a>}}
\if{latex}{\out{\hypertarget{method-Parser-p_simple_if_statement}{}}}
\subsection{\code{Parser$p_simple_if_statement()}}{
  function to handle simple_if_statement
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Parser$p_simple_if_statement(
  doc = "simple_if_statement : IF CONDITION EQUATION",
  p
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{doc}}{doc argument}
      \item{\code{p}}{parser object}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parser-p_if_statement"></a>}}
\if{latex}{\out{\hypertarget{method-Parser-p_if_statement}{}}}
\subsection{\code{Parser$p_if_statement()}}{
  function to handle if_statement
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Parser$p_if_statement(
  doc =
    "if_statement : IF CONDITION EQUATION\\n                            | IF CONDITION IF_CONTENT\\n                            | ELSE_IF CONDITION IF_CONTENT\\n                            | ELSE IF_CONTENT",
  p
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{doc}}{doc argument}
      \item{\code{p}}{parser object}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parser-p_equation"></a>}}
\if{latex}{\out{\hypertarget{method-Parser-p_equation}{}}}
\subsection{\code{Parser$p_equation()}}{
  function to handle equation
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Parser$p_equation(doc = "equation : EQUATION", p)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{doc}}{doc argument}
      \item{\code{p}}{parser object}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parser-p_ode"></a>}}
\if{latex}{\out{\hypertarget{method-Parser-p_ode}{}}}
\subsection{\code{Parser$p_ode()}}{
  function to handle ode
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Parser$p_ode(doc = "ode : ODE", p)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{doc}}{doc argument}
      \item{\code{p}}{parser object}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parser-p_line_break"></a>}}
\if{latex}{\out{\hypertarget{method-Parser-p_line_break}{}}}
\subsection{\code{Parser$p_line_break()}}{
  function to handle line_break
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Parser$p_line_break(doc = "line_break : LINE_BREAK", p)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{doc}}{doc argument}
      \item{\code{p}}{parser object}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parser-p_comment"></a>}}
\if{latex}{\out{\hypertarget{method-Parser-p_comment}{}}}
\subsection{\code{Parser$p_comment()}}{
  function to handle comment
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Parser$p_comment(doc = "comment : COMMENT", p)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{doc}}{doc argument}
      \item{\code{p}}{parser object}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parser-p_unknown_statement"></a>}}
\if{latex}{\out{\hypertarget{method-Parser-p_unknown_statement}{}}}
\subsection{\code{Parser$p_unknown_statement()}}{
  function to handle unknown_statement
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Parser$p_unknown_statement(doc = "unknown_statement : UNKNOWN_STATEMENT", p)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{doc}}{doc argument}
      \item{\code{p}}{parser object}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parser-p_error"></a>}}
\if{latex}{\out{\hypertarget{method-Parser-p_error}{}}}
\subsection{\code{Parser$p_error()}}{
  function to handle errors
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Parser$p_error(p)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{p}}{parser object}
      \item{\code{doc}}{doc argument}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parser-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Parser-clone}{}}}
\subsection{\code{Parser$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Parser$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
