% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxode_conversion.R
\name{rxodeOmega}
\alias{rxodeOmega}
\title{Get the IIV matrix (omega) for RxODE.}
\usage{
rxodeOmega(model, estimate = TRUE)
}
\arguments{
\item{model}{PMX model}

\item{estimate}{if TRUE, estimated values are used, if FALSE, initial values are used}
}
\value{
named matrix with OMEGA values
}
\description{
Get the IIV matrix (omega) for RxODE.
}
