% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxode_conversion.R
\name{rxodeSigma}
\alias{rxodeSigma}
\title{Get the RUV matrix (sigma) for RxODE.
No correlations are possible for now.}
\usage{
rxodeSigma(model, estimate = TRUE)
}
\arguments{
\item{model}{PMX model}

\item{estimate}{if TRUE, estimated values are used, if FALSE, initial values are used}
}
\value{
named matrix with SIGMA values
}
\description{
Get the RUV matrix (sigma) for RxODE.
No correlations are possible for now.
}
