% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxode_conversion.R
\name{rxodeTheta}
\alias{rxodeTheta}
\title{Get the THETA vector for RxODE.}
\usage{
rxodeTheta(model, estimate = TRUE)
}
\arguments{
\item{model}{PMX model}

\item{estimate}{if TRUE, estimated values are used, if FALSE, initial values are used}
}
\value{
named vector with THETA values
}
\description{
Get the THETA vector for RxODE.
}
