% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_mapping.R
\name{searchCandidateName}
\alias{searchCandidateName}
\title{Seach candidate name.}
\usage{
searchCandidateName(model, parameter)
}
\arguments{
\item{model}{CAMPSIS model}

\item{parameter}{parameter to search for a name in model}
}
\value{
a candidate name or NA if nothing was found
}
\description{
Seach candidate name.
}
