% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonmem_qualification_prep.R
\name{setEtasAsCovariates}
\alias{setEtasAsCovariates}
\title{Set ETAs as covariates in NONMEM block.}
\usage{
setEtasAsCovariates(model, content)
}
\arguments{
\item{model}{Campsis parameters}

\item{content}{NONMEM content}
}
\value{
updated content with ETAs replaced by their names
}
\description{
Set ETAs as covariates in NONMEM block.
}
