% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonmem_qualification_prep.R
\name{standardiseDataset}
\alias{standardiseDataset}
\title{Standardise dataset (dataframe) received from CAMPSIS.
This process will be implemented later on in CAMPSIS (using export dest='NONMEM').}
\usage{
standardiseDataset(dataset)
}
\arguments{
\item{dataset}{dataframe or tibble}
}
\value{
a standardised dataset
}
\description{
Standardise dataset (dataframe) received from CAMPSIS.
This process will be implemented later on in CAMPSIS (using export dest='NONMEM').
}
