% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonmem_dataset_parser.R
\name{standardiseNMDataset}
\alias{standardiseNMDataset}
\title{Standardise NONMEM dataset.}
\usage{
standardiseNMDataset(dataset)
}
\arguments{
\item{dataset}{NONMEM dataset}
}
\value{
updated dataset
}
\description{
Standardise NONMEM dataset.
}
