% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmx_model_conversion.R
\name{statementToPmxModel}
\alias{statementToPmxModel}
\title{SymPy statement conversion to PMX model.}
\usage{
statementToPmxModel(statement, params)
}
\arguments{
\item{statement}{SymPy statement}

\item{params}{parameters definition table}
}
\value{
C code
}
\description{
SymPy statement conversion to PMX model.
}
