% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxode_conversion.R
\name{statementToRxODE}
\alias{statementToRxODE}
\title{SymPy statement conversion to RxODE code.}
\usage{
statementToRxODE(statement, params)
}
\arguments{
\item{statement}{SymPy statement}

\item{params}{parameters definition table}
}
\value{
RxODE code
}
\description{
SymPy statement conversion to RxODE code.
}
