% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substitute_duplicates.R
\name{substituteDuplicateEquationNames}
\alias{substituteDuplicateEquationNames}
\title{Substitute duplicate equation names (using underscore characters).}
\usage{
substituteDuplicateEquationNames(model)
}
\arguments{
\item{model}{Campsis model with duplicate equation names}
}
\value{
Campsis model without duplicate equation names
}
\description{
Substitute duplicate equation names (using underscore characters).
}
