% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substitute_duplicates.R
\name{substituteDuplicates}
\alias{substituteDuplicates}
\title{Substitute duplicate equations (using underscore characters).}
\usage{
substituteDuplicates(model)
}
\description{
Substitute duplicate equations (using underscore characters).
}
