% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbol_conversion.R
\name{toDiagonalMatrix}
\alias{toDiagonalMatrix}
\title{Diagonal vector to matrix utility method.}
\usage{
toDiagonalMatrix(diagVector)
}
\arguments{
\item{diagVector}{the diagonal vector}
}
\value{
a matrix
}
\description{
Diagonal vector to matrix utility method.
}
