% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonmem_update.R
\name{toNONMEMPop}
\alias{toNONMEMPop}
\title{Export PMXtran object to NONMEM control stream for population simulation (POP).
If NONMEM results are provided, they will replace the previous original values in the control stream.
OMEGA and SIGMA initial values will then be fixed and set to 0.
ETA arrays in control stream will be replaced by ETA covariates (e.g. ETA(1) -> ETA_1)}
\usage{
toNONMEMPop(pmxtran)
}
\arguments{
\item{pmxtran}{PMXtran object}
}
\value{
the updated PMXtran object
}
\description{
Export PMXtran object to NONMEM control stream for population simulation (POP).
If NONMEM results are provided, they will replace the previous original values in the control stream.
OMEGA and SIGMA initial values will then be fixed and set to 0.
ETA arrays in control stream will be replaced by ETA covariates (e.g. ETA(1) -> ETA_1)
}
