% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonmem_update.R
\name{toNONMEMPred}
\alias{toNONMEMPred}
\title{Export PMX model to NONMEM control stream for typical profile simulation (PRED).
If NONMEM results are provided, they will replace the previous original values in the control stream.
OMEGA and SIGMA initial values will then be fixed and set to 0.}
\usage{
toNONMEMPred(model)
}
\arguments{
\item{model}{PMX model}
}
\value{
the updated PMX model
}
\description{
Export PMX model to NONMEM control stream for typical profile simulation (PRED).
If NONMEM results are provided, they will replace the previous original values in the control stream.
OMEGA and SIGMA initial values will then be fixed and set to 0.
}
