% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharmpy_domain.R
\name{toNONMEMStyle}
\alias{toNONMEMStyle}
\title{Get type of NONMEM parameter.}
\usage{
toNONMEMStyle(str)
}
\arguments{
\item{str}{string}
}
\value{
the type of NONMEM parameter
}
\description{
Get type of NONMEM parameter.
}
