% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmx_model_conversion.R
\name{toPmxModel}
\alias{toPmxModel}
\title{PMXtran model conversion to PMX model.}
\usage{
toPmxModel(pmxtran)
}
\arguments{
\item{pmxtran}{PMXtran model}
}
\value{
PMX model
}
\description{
PMXtran model conversion to PMX model.
}
