% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxode_conversion.R
\name{toRxODE}
\alias{toRxODE}
\title{Pharmpy model conversion to RxODE code.}
\usage{
toRxODE(model)
}
\arguments{
\item{model}{PMX model}
}
\value{
RxODE code
}
\description{
Pharmpy model conversion to RxODE code.
}
