% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonmem_qualification_prep.R
\name{updateControlStreamForSimulation}
\alias{updateControlStreamForSimulation}
\title{Update control stream for simulation.}
\usage{
updateControlStreamForSimulation(
  model_path,
  estimate = TRUE,
  dataset,
  output_folder,
  variables,
  compartments = NULL
)
}
\arguments{
\item{model_path}{path to original control stream}

\item{estimate}{reuse estimated parameters from the model fit, default is TRUE}

\item{dataset}{simulation dataset, data frame}

\item{output_folder}{output folder to export the qualification control stream and CSV dataset}

\item{variables}{variables to output (note: ID, ARM, TIME, EVID, MDV, DV, AMT, CMT, DOSENO are output by default)}

\item{compartments}{compartment indexes to output, numeric vector}
}
\value{
the updated campsistrans object
}
\description{
Update control stream for simulation.
}
