% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonmem_qualification_prep.R
\name{updateCtlForQual}
\alias{updateCtlForQual}
\title{Update control stream for qualification}
\usage{
updateCtlForQual(model, dataset, campsis, variables, compartments = NULL)
}
\arguments{
\item{model}{Pharmpy model}

\item{dataset}{simulation dataset, data frame}

\item{campsis}{Campsis model, used for mapping ETAs to covariates}

\item{variables}{variables to output (note: ID, ARM, TIME, EVID, MDV, DV, AMT, CMT, DOSENO are output by default)}

\item{compartments}{compartment indexes to output, numeric vector}
}
\value{
the updated campsistrans object
}
\description{
Update control stream for qualification
}
