% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{write.default}
\alias{write.default}
\title{Default write function.}
\usage{
\method{write}{default}(x, file, ...)
}
\arguments{
\item{x}{object to write}

\item{file}{destination file}

\item{...}{ignored}
}
\description{
Default write function.
}
