% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmxtran.R
\name{write.pmxtran}
\alias{write.pmxtran}
\title{Write PMXtran object to NONMEM control stream file.}
\usage{
write.pmxtran(x, file, ...)
}
\arguments{
\item{x}{PMXtran object}

\item{file}{path to exported NONMEM control stream}

\item{...}{ignored}
}
\description{
Write PMXtran object to NONMEM control stream file.
}
