% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{write}
\alias{write}
\title{Write is a generic function to write models.}
\usage{
write(x, file, ...)
}
\arguments{
\item{x}{object to write}

\item{file}{destination file}

\item{...}{ignored}
}
\description{
Write is a generic function to write models.
}
