% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonmem_qualification_prep.R
\name{writeCtl}
\alias{writeCtl}
\title{Write control stream to files.}
\usage{
writeCtl(model, path, force)
}
\arguments{
\item{model}{Pharmpy model}

\item{path}{where to write the control stream}

\item{force}{if TRUE, overwrite existing file}
}
\value{
the path to the written control stream
}
\description{
Write control stream to files.
}
