% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonmem_qualification_prep.R
\name{writeSimulationCtl}
\alias{writeSimulationCtl}
\title{Write simulation control stream.}
\usage{
writeSimulationCtl(campsistrans, folder)
}
\arguments{
\item{campsistrans}{Campsistrans object}

\item{folder}{the path to the qualification folder}
}
\value{
the path to the qualification folder
}
\description{
Write simulation control stream.
}
