% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonmem_dataset_parser.R
\name{addSimulationIDColumn}
\alias{addSimulationIDColumn}
\title{Add simulation ID column.}
\usage{
addSimulationIDColumn(dataset, id = "ID")
}
\arguments{
\item{dataset}{NONMEM dataset}

\item{id}{}
}
\value{
updated data frame
}
\description{
Add simulation ID column.
}
