% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_mapping.R
\name{autoRenameParameters}
\alias{autoRenameParameters}
\title{Auto-rename parameters based on equations.}
\usage{
autoRenameParameters(model)
}
\arguments{
\item{model}{Campsis model}
}
\value{
Campsis model with updated parameters
}
\description{
Auto-rename parameters based on equations.
}
