% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret_to_pow.R
\name{caretToPow}
\alias{caretToPow}
\alias{caretToPow,model_statement-method}
\alias{caretToPow,character-method}
\alias{caretToPow,equation-method}
\alias{caretToPow,if_statement-method}
\alias{caretToPow,code_record-method}
\alias{caretToPow,code_records-method}
\alias{caretToPow,compartment_property-method}
\alias{caretToPow,campsis_model-method}
\title{Convert the caret operator into the pow function.}
\usage{
caretToPow(x)

caretToPow(x)

\S4method{caretToPow}{model_statement}(x)

\S4method{caretToPow}{character}(x)

\S4method{caretToPow}{equation}(x)

\S4method{caretToPow}{if_statement}(x)

\S4method{caretToPow}{code_record}(x)

\S4method{caretToPow}{code_records}(x)

\S4method{caretToPow}{compartment_property}(x)

\S4method{caretToPow}{campsis_model}(x)
}
\arguments{
\item{x}{object to transform}
}
\description{
Convert the caret operator into the pow function.
}
