% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_conversion.R
\name{convertCompartmentSystem}
\alias{convertCompartmentSystem}
\title{Pharmpy compartment system conversion to PMX model.}
\usage{
convertCompartmentSystem(system)
}
\arguments{
\item{system}{Pharmpy compartment system}
}
\value{
DES record (PMX domain)
}
\description{
Pharmpy compartment system conversion to PMX model.
}
