% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varcov_conversion.R
\name{convertNONMEMParameter}
\alias{convertNONMEMParameter}
\title{Convert NONMEM parameter (string form) to campsismod parameter.}
\usage{
convertNONMEMParameter(name, value, fix)
}
\arguments{
\item{name}{NONMEM parameter name, character value}

\item{value}{parameter value}

\item{fix}{is fixed or not, logical value}
}
\value{
S4 parameters object
}
\description{
Convert NONMEM parameter (string form) to campsismod parameter.
}
