% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_conversion.R
\name{convertPiecewise}
\alias{convertPiecewise}
\title{SymPy piecewise conversion to CAMPSIS model.}
\usage{
convertPiecewise(symbol, expression)
}
\arguments{
\item{symbol}{SymPy statement symbol}

\item{expression}{SymPy piecewise}
}
\value{
C code
}
\description{
SymPy piecewise conversion to CAMPSIS model.
}
