% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex_if_statement.R
\name{ElseIfStatement}
\alias{ElseIfStatement}
\title{Create a new ELSE-IF statement.}
\usage{
ElseIfStatement(condition, statements)
}
\arguments{
\item{condition}{condition, single character string}

\item{statements}{embedded statements}
}
\value{
an ELSE-IF statement
}
\description{
Create a new ELSE-IF statement.
}
