% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex_if_statement.R
\name{ElseStatement}
\alias{ElseStatement}
\title{Create a new ELSE statement.}
\usage{
ElseStatement(equation, comment = as.character(NA))
}
\arguments{
\item{equation}{equation if condition is met}

\item{comment}{comment if any, single character string}
}
\value{
an ELSE statement
}
\description{
Create a new ELSE statement.
}
