% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonmem_qualification_prep.R
\name{executeSimulationCtl}
\alias{executeSimulationCtl}
\title{Prepare simulation control stream and execute it with NONMEM.}
\usage{
executeSimulationCtl(
  campsistrans,
  dataset,
  variables,
  compartments = NULL,
  folder,
  reexecuteNONMEM = T
)
}
\arguments{
\item{campsistrans}{campsistrans object}

\item{dataset}{simulation dataset, data frame}

\item{variables}{variables to output (note: ID, ARM, TIME, EVID, MDV, DV, AMT, CMT, DOSENO are output by default)}

\item{compartments}{compartment indexes to output, numeric vector}

\item{folder}{where to execute the simulation control stream}

\item{reexecuteNONMEM}{force re-execute NONMEM if results already exist}
}
\value{
a data frame with NONMEM results
}
\description{
Prepare simulation control stream and execute it with NONMEM.
}
