% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_from_rxode2.R
\name{extractParametersFromRxode}
\alias{extractParametersFromRxode}
\title{Extract all parameters from an rxode2 function object.}
\usage{
extractParametersFromRxode(rxmod, rem_pop_suffix, rem_omega_prefix)
}
\arguments{
\item{rxmod}{the rxode2 function object}

\item{rem_pop_suffix}{remove the '_pop' suffix from the population parameters.}

\item{rem_omega_prefix}{remove the 'omega_' prefix from the OMEGA parameters.}
}
\value{
a functional Campsis model
}
\description{
Extract all parameters from an rxode2 function object.
}
