% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonmem_dataset_parser.R
\name{importDataset}
\alias{importDataset}
\title{Import a NONMEM dataset and prepare it for the qualification.}
\usage{
importDataset(
  file,
  covariates = NULL,
  etas = FALSE,
  table_no = NULL,
  etas_zero = FALSE,
  campsis = NULL,
  campsis_id = FALSE
)
}
\arguments{
\item{file}{path to NONMEM control stream (dataset path will be automatically extracted)}

\item{covariates}{covariates vector. If provided, only these covariates are kept in dataset.
NULL is default (all covariates are kept)}

\item{etas}{import estimated ETAs that were output by $TABLE in control stream.
By default, ETAs are not imported. Please set it to TRUE to enable this feature.}

\item{table_no}{table number to look for the ETAs, if NULL, first table with ETAS(1:LAST) occurrence will be used}

\item{etas_zero}{if previous argument is set to FALSE, etas_zero set to TRUE will
all ETAs from model to 0 (useful to simulate model without IIV)}

\item{campsis}{Campsis model object, if provided, ETAs in the dataset will be renamed according to the Campsis model}

\item{campsis_id}{rework ID column for simulation with Campsis (ID must start at 1 and must be consecutive).
Default is FALSE. Please set it to TRUE if you wish a simulation ID. If TRUE, original ID column is
preserved in column 'ORIGINAL_ID'.}
}
\value{
a data frame
}
\description{
Import a NONMEM dataset and prepare it for the qualification.
}
