% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_from_monolix.R
\name{importMonolix}
\alias{importMonolix}
\title{Extract the model from Monolix.}
\usage{
importMonolix(
  mlxtranFile,
  modelFile = NULL,
  parametersFile = NULL,
  covFile = NULL,
  digits = NULL
)
}
\arguments{
\item{mlxtranFile}{to mlxtran file}

\item{modelFile}{path to model file, optional}

\item{parametersFile}{path to estimated parameters file, optional}

\item{covFile}{path to covariance estimates file, optional}

\item{digits}{number of significant digits, integer or NULL to disable rounding}
}
\value{
a functional Campsis model
}
\description{
Extract the model from Monolix.
}
