% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmxtran.R
\name{importNONMEM}
\alias{importNONMEM}
\title{Import NONMEM control stream and results using Pharmpy.}
\usage{
importNONMEM(file, mapping = NULL, estimate = FALSE, uncertainty = FALSE)
}
\arguments{
\item{file}{path to NONMEM control stream, character value}

\item{mapping}{a possible PMX mapping object}

\item{estimate}{use estimated parameter values or initial ones, logical value, default is FALSE}

\item{uncertainty}{import the variance-covariance matrix (.cov file), logical value, default is FALSE}
}
\value{
a PMXtran object
}
\description{
Import NONMEM control stream and results using Pharmpy.
}
