% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campsistrans.R
\name{importNONMEM}
\alias{importNONMEM}
\title{Import NONMEM control stream and results using Pharmpy.}
\usage{
importNONMEM(
  file,
  mapping = NULL,
  estimate = FALSE,
  uncertainty = FALSE,
  covar_name = FALSE,
  covar_as_cor = FALSE,
  auto_install = TRUE,
  pharmpy_config = OldPharmpyConfig(),
  copy_dir = FALSE,
  rem_rate = FALSE,
  rem_abbr_replace = TRUE
)
}
\arguments{
\item{file}{path to NONMEM control stream, character value}

\item{mapping}{a possible PMX mapping object}

\item{estimate}{use estimated parameter values or initial ones, logical value, default is FALSE}

\item{uncertainty}{import the variance-covariance matrix (.cov file), logical value, default is FALSE}

\item{covar_name}{give a name to each covariance value}

\item{covar_as_cor}{transform covariance values to correlation values}

\item{auto_install}{auto install pharmpy and dependencies if not installed yet}

\item{pharmpy_config}{a Pharmpy configuration object, see \link{OldPharmpyConfig} for details}

\item{copy_dir}{copy directory in which the control stream is}

\item{rem_rate}{remove RATE in control stream automatically to avoid issues with Pharmpy.
Otherwise, it will look for the dataset and possibly adapt the ODE's to add the rates, default is FALSE}

\item{rem_abbr_replace}{remove section ABBREVIATED REPLACE, causing issue in import, default is TRUE}
}
\value{
a campsistrans object
}
\description{
Import NONMEM control stream and results using Pharmpy.
}
