% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campsistrans.R
\name{importNONMEM}
\alias{importNONMEM}
\title{Import NONMEM control stream and results using Pharmpy.}
\usage{
importNONMEM(
  file,
  mapping = NULL,
  estimate = FALSE,
  uncertainty = FALSE,
  auto_install = TRUE,
  envname = getPythonEnvName(),
  python = getPythonPath()
)
}
\arguments{
\item{file}{path to NONMEM control stream, character value}

\item{mapping}{a possible PMX mapping object}

\item{estimate}{use estimated parameter values or initial ones, logical value, default is FALSE}

\item{uncertainty}{import the variance-covariance matrix (.cov file), logical value, default is FALSE}

\item{auto_install}{auto install pharmpy and dependencies if not installed yet}

\item{envname}{virtual python environment name, can be configured in config.yml}

\item{python}{path to python, can be configured in config.yml}
}
\value{
a campsistrans object
}
\description{
Import NONMEM control stream and results using Pharmpy.
}
