% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharmpy_config.R
\name{importPharmpyPackage}
\alias{importPharmpyPackage}
\title{Import python package without error.}
\usage{
importPharmpyPackage(config = UpdatedPharmpyConfig())
}
\arguments{
\item{config}{pharmpy config, default is UpdatedPharmpyConfig()}
}
\value{
NULL if an error occurs
}
\description{
Import python package without error.
}
