% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_from_rxode2.R
\name{importRxode2}
\alias{importRxode2}
\title{Extract model code from an rxode2 function object.}
\usage{
importRxode2(rxModel)
}
\arguments{
\item{rxModel}{the rxode2 function object}
}
\value{
a functional Campsis model
}
\description{
Extract model code from an rxode2 function object.
}
