% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_from_rxode2.R
\name{importRxode2}
\alias{importRxode2}
\title{Extract model code from an rxode2 function object.}
\usage{
importRxode2(rxmod, rem_pop_str = FALSE, rem_omega_str = FALSE)
}
\arguments{
\item{rxmod}{the rxode2 function object}

\item{rem_pop_suffix}{remove the '_pop' suffix from the population parameters. The model code is adapted accordingly.}

\item{rem_omega_prefix}{remove the 'omega_' prefix from the OMEGA parameters. The model code is adapted accordingly.}
}
\value{
a functional Campsis model
}
\description{
Extract model code from an rxode2 function object.
}
