% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_conversion.R
\name{mapping}
\alias{mapping}
\title{Create PMX mapping.}
\usage{
mapping(theta = NULL, omega = NULL, sigma = NULL, auto = FALSE)
}
\arguments{
\item{theta}{named integer vector for THETA mapping}

\item{omega}{named integer vector for THETA mapping}

\item{sigma}{named integer vector for OMEGA mapping}

\item{auto}{derive parameter names based on NONMEM equations. Default value
is FALSE for backward compatibility.}
}
\value{
PMX mapping object
}
\description{
Create PMX mapping.
}
