% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmxtran.R
\name{modelToNONMEM}
\alias{modelToNONMEM}
\title{Write PMXtran object to NONMEM control stream file.}
\usage{
modelToNONMEM(pmxtran, path)
}
\arguments{
\item{path}{path to desired control stream file}

\item{model}{PMX model}
}
\description{
Write PMXtran object to NONMEM control stream file.
}
