% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbol_conversion.R
\name{nameParameter}
\alias{nameParameter}
\title{Name the given parameter type according to the parameters table.}
\usage{
nameParameter(type, params)
}
\arguments{
\item{type}{parameter type}

\item{params}{parameters}
}
\value{
a pretty parameter name
}
\description{
Name the given parameter type according to the parameters table.
}
