% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmxtran.R
\name{nameParameter}
\alias{nameParameter}
\title{Name the given parameter type according to the parameters definition table.}
\usage{
nameParameter(type, params)
}
\arguments{
\item{type}{parameter type}

\item{params}{parameters definition table}
}
\value{
a pretty parameter name
}
\description{
Name the given parameter type according to the parameters definition table.
}
