% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonmem_qualification_prep.R
\name{omegaSigmaToZero}
\alias{omegaSigmaToZero}
\title{Set OMEGA and SIGMA initial values to 0 and fix them.}
\usage{
omegaSigmaToZero(x)
}
\arguments{
\item{x}{campsistrans object}
}
\value{
the updated campsistrans object
}
\description{
Set OMEGA and SIGMA initial values to 0 and fix them.
}
