% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonmem_qualification_prep.R
\name{prepareSimulationCtl}
\alias{prepareSimulationCtl}
\title{Prepare simulation control stream.}
\usage{
prepareSimulationCtl(campsistrans, dataset, variables)
}
\arguments{
\item{campsistrans}{Pharmpy model}

\item{dataset}{simulation dataset, data frame}

\item{variables}{variables to output}

\item{campsis}{Campsis model, used for mapping ETAs to covariates}
}
\value{
the updated Campsistrans object
}
\description{
Prepare simulation control stream.
}
