% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmx_model_conversion.R
\name{recordToPmxModel}
\alias{recordToPmxModel}
\title{NONMEM record (pharmpy) to PMX model.}
\usage{
recordToPmxModel(records, recordType, parameters)
}
\arguments{
\item{records}{one or more NONMEM record}

\item{recordType}{type of record}

\item{parameters}{parameters}
}
\value{
a PMX record
}
\description{
NONMEM record (pharmpy) to PMX model.
}
