% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_model_rxode2_parser.R
\name{Rxode2ErrorModelParser}
\alias{Rxode2ErrorModelParser}
\title{Rxode2 error model parser.}
\value{
a parser object
}
\description{
Rxode2ErrorModelParser class to handle the parser for Rxode2 error model
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{tokens}}{list of tokens}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Parser-p_merged_error_model}{\code{Rxode2ErrorModelParser$p_merged_error_model()}}
\item \href{#method-Parser-p_error_model}{\code{Rxode2ErrorModelParser$p_error_model()}}
\item \href{#method-Parser-p_endpoint}{\code{Rxode2ErrorModelParser$p_endpoint()}}
\item \href{#method-Parser-p_add}{\code{Rxode2ErrorModelParser$p_add()}}
\item \href{#method-Parser-p_prop}{\code{Rxode2ErrorModelParser$p_prop()}}
\item \href{#method-Parser-p_combined1}{\code{Rxode2ErrorModelParser$p_combined1()}}
\item \href{#method-Parser-p_combined2}{\code{Rxode2ErrorModelParser$p_combined2()}}
\item \href{#method-Parser-p_error}{\code{Rxode2ErrorModelParser$p_error()}}
\item \href{#method-Parser-clone}{\code{Rxode2ErrorModelParser$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parser-p_merged_error_model"></a>}}
\if{latex}{\out{\hypertarget{method-Parser-p_merged_error_model}{}}}
\subsection{Method \code{p_merged_error_model()}}{
tokens list of tokens


function to handle merged error model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rxode2ErrorModelParser$p_merged_error_model(
  doc =
    "merged_error_model : error_model\\n                         | merged_error_model error_model",
  p
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{doc}}{doc argument}

\item{\code{p}}{parser object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parser-p_error_model"></a>}}
\if{latex}{\out{\hypertarget{method-Parser-p_error_model}{}}}
\subsection{Method \code{p_error_model()}}{
function to handle error model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rxode2ErrorModelParser$p_error_model(
  doc =
    "error_model : endpoint\\n                         | add\\n                         | prop\\n                         | combined1\\n                         | combined2",
  p
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{doc}}{doc argument}

\item{\code{p}}{parser object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parser-p_endpoint"></a>}}
\if{latex}{\out{\hypertarget{method-Parser-p_endpoint}{}}}
\subsection{Method \code{p_endpoint()}}{
function to handle endpoint
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rxode2ErrorModelParser$p_endpoint(doc = "endpoint : ENDPOINT", p)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{doc}}{doc argument}

\item{\code{p}}{parser object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parser-p_add"></a>}}
\if{latex}{\out{\hypertarget{method-Parser-p_add}{}}}
\subsection{Method \code{p_add()}}{
function to handle add
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rxode2ErrorModelParser$p_add(doc = "add : ADD", p)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{doc}}{doc argument}

\item{\code{p}}{parser object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parser-p_prop"></a>}}
\if{latex}{\out{\hypertarget{method-Parser-p_prop}{}}}
\subsection{Method \code{p_prop()}}{
function to handle prop
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rxode2ErrorModelParser$p_prop(doc = "prop : PROP", p)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{doc}}{doc argument}

\item{\code{p}}{parser object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parser-p_combined1"></a>}}
\if{latex}{\out{\hypertarget{method-Parser-p_combined1}{}}}
\subsection{Method \code{p_combined1()}}{
function to handle combined1
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rxode2ErrorModelParser$p_combined1(doc = "combined1 : COMBINED1", p)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{doc}}{doc argument}

\item{\code{p}}{parser object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parser-p_combined2"></a>}}
\if{latex}{\out{\hypertarget{method-Parser-p_combined2}{}}}
\subsection{Method \code{p_combined2()}}{
function to handle combined2
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rxode2ErrorModelParser$p_combined2(doc = "combined2 : COMBINED2", p)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{doc}}{doc argument}

\item{\code{p}}{parser object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parser-p_error"></a>}}
\if{latex}{\out{\hypertarget{method-Parser-p_error}{}}}
\subsection{Method \code{p_error()}}{
function to handle errors
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rxode2ErrorModelParser$p_error(p)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{p}}{parser object}

\item{\code{doc}}{doc argument}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parser-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Parser-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rxode2ErrorModelParser$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
