% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxode2_parser.R
\name{Rxode2Lexer}
\alias{Rxode2Lexer}
\title{Rxode2 lexer.}
\description{
Rxode2 lexer.

Rxode2 lexer.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Lexer-t_EQUATION}{\code{Rxode2Lexer$t_EQUATION()}}
\item \href{#method-Lexer-t_ODE}{\code{Rxode2Lexer$t_ODE()}}
\item \href{#method-Lexer-t_LINE_BREAK}{\code{Rxode2Lexer$t_LINE_BREAK()}}
\item \href{#method-Lexer-t_COMMENT}{\code{Rxode2Lexer$t_COMMENT()}}
\item \href{#method-Lexer-t_IF}{\code{Rxode2Lexer$t_IF()}}
\item \href{#method-Lexer-t_ELSE_IF}{\code{Rxode2Lexer$t_ELSE_IF()}}
\item \href{#method-Lexer-t_ELSE}{\code{Rxode2Lexer$t_ELSE()}}
\item \href{#method-Lexer-t_CONDITION}{\code{Rxode2Lexer$t_CONDITION()}}
\item \href{#method-Lexer-t_IF_CONTENT}{\code{Rxode2Lexer$t_IF_CONTENT()}}
\item \href{#method-Lexer-t_UNKNOWN_STATEMENT}{\code{Rxode2Lexer$t_UNKNOWN_STATEMENT()}}
\item \href{#method-Lexer-t_newline}{\code{Rxode2Lexer$t_newline()}}
\item \href{#method-Lexer-t_error}{\code{Rxode2Lexer$t_error()}}
\item \href{#method-Lexer-clone}{\code{Rxode2Lexer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lexer-t_EQUATION"></a>}}
\if{latex}{\out{\hypertarget{method-Lexer-t_EQUATION}{}}}
\subsection{Method \code{t_EQUATION()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rxode2Lexer$t_EQUATION(
  re = "[a-zA-Z_][a-zA-Z0-9_\\\\.]*\\\\s*=\\\\s*[^\\n]+",
  t
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lexer-t_ODE"></a>}}
\if{latex}{\out{\hypertarget{method-Lexer-t_ODE}{}}}
\subsection{Method \code{t_ODE()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rxode2Lexer$t_ODE(re = "d/dt\\\\s*\\\\(.*\\\\)\\\\s*=\\\\s*[^\\n]+", t)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lexer-t_LINE_BREAK"></a>}}
\if{latex}{\out{\hypertarget{method-Lexer-t_LINE_BREAK}{}}}
\subsection{Method \code{t_LINE_BREAK()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rxode2Lexer$t_LINE_BREAK(re = "\\n[ ]*\\n", t)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lexer-t_COMMENT"></a>}}
\if{latex}{\out{\hypertarget{method-Lexer-t_COMMENT}{}}}
\subsection{Method \code{t_COMMENT()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rxode2Lexer$t_COMMENT(re = "\\\\s*#.*[^\\n]+", t)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lexer-t_IF"></a>}}
\if{latex}{\out{\hypertarget{method-Lexer-t_IF}{}}}
\subsection{Method \code{t_IF()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rxode2Lexer$t_IF(re = "if", t)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lexer-t_ELSE_IF"></a>}}
\if{latex}{\out{\hypertarget{method-Lexer-t_ELSE_IF}{}}}
\subsection{Method \code{t_ELSE_IF()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rxode2Lexer$t_ELSE_IF(re = "else\\\\s+if", t)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lexer-t_ELSE"></a>}}
\if{latex}{\out{\hypertarget{method-Lexer-t_ELSE}{}}}
\subsection{Method \code{t_ELSE()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rxode2Lexer$t_ELSE(re = "else", t)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lexer-t_CONDITION"></a>}}
\if{latex}{\out{\hypertarget{method-Lexer-t_CONDITION}{}}}
\subsection{Method \code{t_CONDITION()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rxode2Lexer$t_CONDITION(re = "\\\\((?>[^()]|(?R))*\\\\)", t)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lexer-t_IF_CONTENT"></a>}}
\if{latex}{\out{\hypertarget{method-Lexer-t_IF_CONTENT}{}}}
\subsection{Method \code{t_IF_CONTENT()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rxode2Lexer$t_IF_CONTENT(re = "\\\\{(?>[^{}]|(?R))*\\\\}", t)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lexer-t_UNKNOWN_STATEMENT"></a>}}
\if{latex}{\out{\hypertarget{method-Lexer-t_UNKNOWN_STATEMENT}{}}}
\subsection{Method \code{t_UNKNOWN_STATEMENT()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rxode2Lexer$t_UNKNOWN_STATEMENT(re = ".*[^\\n]+", t)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lexer-t_newline"></a>}}
\if{latex}{\out{\hypertarget{method-Lexer-t_newline}{}}}
\subsection{Method \code{t_newline()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rxode2Lexer$t_newline(re = "\\\\n+", t)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lexer-t_error"></a>}}
\if{latex}{\out{\hypertarget{method-Lexer-t_error}{}}}
\subsection{Method \code{t_error()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rxode2Lexer$t_error(t)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lexer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Lexer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rxode2Lexer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
