% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxode2_parser.R
\name{Rxode2Parser}
\alias{Rxode2Parser}
\title{Rxode2 parser.}
\description{
Rxode2 parser.

Rxode2 parser.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Parser-p_model_statements}{\code{Rxode2Parser$p_model_statements()}}
\item \href{#method-Parser-p_model_statement}{\code{Rxode2Parser$p_model_statement()}}
\item \href{#method-Parser-p_complex_if_else_statement}{\code{Rxode2Parser$p_complex_if_else_statement()}}
\item \href{#method-Parser-p_simple_if_statement}{\code{Rxode2Parser$p_simple_if_statement()}}
\item \href{#method-Parser-p_if_statement}{\code{Rxode2Parser$p_if_statement()}}
\item \href{#method-Parser-p_equation}{\code{Rxode2Parser$p_equation()}}
\item \href{#method-Parser-p_ode}{\code{Rxode2Parser$p_ode()}}
\item \href{#method-Parser-p_line_break}{\code{Rxode2Parser$p_line_break()}}
\item \href{#method-Parser-p_comment}{\code{Rxode2Parser$p_comment()}}
\item \href{#method-Parser-p_unknown_statement}{\code{Rxode2Parser$p_unknown_statement()}}
\item \href{#method-Parser-p_error}{\code{Rxode2Parser$p_error()}}
\item \href{#method-Parser-clone}{\code{Rxode2Parser$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parser-p_model_statements"></a>}}
\if{latex}{\out{\hypertarget{method-Parser-p_model_statements}{}}}
\subsection{Method \code{p_model_statements()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rxode2Parser$p_model_statements(
  doc =
    "model_statements : model_statement\\n                         | model_statements model_statement",
  p
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parser-p_model_statement"></a>}}
\if{latex}{\out{\hypertarget{method-Parser-p_model_statement}{}}}
\subsection{Method \code{p_model_statement()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rxode2Parser$p_model_statement(
  doc =
    "model_statement : complex_if_else_statement\\n                         | equation \\n                         | ode\\n                         | line_break\\n                         | comment\\n                         | simple_if_statement\\n                         | unknown_statement",
  p
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parser-p_complex_if_else_statement"></a>}}
\if{latex}{\out{\hypertarget{method-Parser-p_complex_if_else_statement}{}}}
\subsection{Method \code{p_complex_if_else_statement()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rxode2Parser$p_complex_if_else_statement(
  doc =
    "complex_if_else_statement : if_statement\\n                         | complex_if_else_statement if_statement",
  p
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parser-p_simple_if_statement"></a>}}
\if{latex}{\out{\hypertarget{method-Parser-p_simple_if_statement}{}}}
\subsection{Method \code{p_simple_if_statement()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rxode2Parser$p_simple_if_statement(
  doc = "simple_if_statement : IF CONDITION EQUATION",
  p
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parser-p_if_statement"></a>}}
\if{latex}{\out{\hypertarget{method-Parser-p_if_statement}{}}}
\subsection{Method \code{p_if_statement()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rxode2Parser$p_if_statement(
  doc =
    "if_statement : IF CONDITION EQUATION\\n                            | IF CONDITION IF_CONTENT\\n                            | ELSE_IF CONDITION IF_CONTENT\\n                            | ELSE IF_CONTENT",
  p
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parser-p_equation"></a>}}
\if{latex}{\out{\hypertarget{method-Parser-p_equation}{}}}
\subsection{Method \code{p_equation()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rxode2Parser$p_equation(doc = "equation : EQUATION", p)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parser-p_ode"></a>}}
\if{latex}{\out{\hypertarget{method-Parser-p_ode}{}}}
\subsection{Method \code{p_ode()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rxode2Parser$p_ode(doc = "ode : ODE", p)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parser-p_line_break"></a>}}
\if{latex}{\out{\hypertarget{method-Parser-p_line_break}{}}}
\subsection{Method \code{p_line_break()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rxode2Parser$p_line_break(doc = "line_break : LINE_BREAK", p)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parser-p_comment"></a>}}
\if{latex}{\out{\hypertarget{method-Parser-p_comment}{}}}
\subsection{Method \code{p_comment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rxode2Parser$p_comment(doc = "comment : COMMENT", p)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parser-p_unknown_statement"></a>}}
\if{latex}{\out{\hypertarget{method-Parser-p_unknown_statement}{}}}
\subsection{Method \code{p_unknown_statement()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rxode2Parser$p_unknown_statement(
  doc = "unknown_statement : UNKNOWN_STATEMENT",
  p
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parser-p_error"></a>}}
\if{latex}{\out{\hypertarget{method-Parser-p_error}{}}}
\subsection{Method \code{p_error()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rxode2Parser$p_error(p)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parser-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Parser-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rxode2Parser$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
