% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxode_conversion.R
\name{rxodeParams}
\alias{rxodeParams}
\title{Access estimate and initial values from model.}
\usage{
rxodeParams(model, nmtype, estimate = TRUE)
}
\arguments{
\item{model}{PMX model}

\item{nmtype}{NONMEM type: THETA, OMEGA or SIGMA}

\item{estimate}{if TRUE, estimated values are used, if FALSE, initial values are used}
}
\value{
named vector with the requested parameters
}
\description{
Access estimate and initial values from model.
}
