% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonmem_qualification_prep.R
\name{setEtasAsCovariates}
\alias{setEtasAsCovariates}
\title{Update ETA's in NONMEM record.}
\usage{
setEtasAsCovariates(statements, params, pharmpy)
}
\arguments{
\item{params}{CAMPSIS parameters}

\item{pharmpy}{pharmpy package}

\item{control_stream}{pharmpy model}
}
\description{
Update ETA's in NONMEM record.
}
