% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonmem_qualification_prep.R
\name{setEtasAsCovariates}
\alias{setEtasAsCovariates}
\title{Set ETAs as covariates in Pharmpy model.}
\usage{
setEtasAsCovariates(statements, params, pharmpy)
}
\arguments{
\item{statements}{pharmpy model}

\item{params}{Campsis parameters}

\item{pharmpy}{pharmpy package}
}
\description{
Set ETAs as covariates in Pharmpy model.
}
