% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_conversion.R
\name{statementToPmxModel}
\alias{statementToPmxModel}
\title{SymPy statement conversion to PMX model.}
\usage{
statementToPmxModel(statement, parameters)
}
\arguments{
\item{statement}{SymPy statement}

\item{parameters}{parameters}
}
\value{
C code
}
\description{
SymPy statement conversion to PMX model.
}
