% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonmem_update.R
\name{updateETAinNONMEMRecord}
\alias{updateETAinNONMEMRecord}
\title{Update ETA's in NONMEM record.}
\usage{
updateETAinNONMEMRecord(ctl, record_type)
}
\arguments{
\item{ctl}{Pharmpy control stream}

\item{record_type}{record type, e.g. "PK"}
}
\description{
Update ETA's in NONMEM record.
}
