% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonmem_qualification_prep.R
\name{updateETAinNONMEMRecord}
\alias{updateETAinNONMEMRecord}
\title{Update ETA's in NONMEM record.}
\usage{
updateETAinNONMEMRecord(control_stream, recordType, params, pharmpy)
}
\arguments{
\item{control_stream}{pharmpy model}

\item{recordType}{record type to adapt}

\item{params}{CAMPSIS parameters}

\item{pharmpy}{pharmpy package}
}
\description{
Update ETA's in NONMEM record.
}
