% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonmem_qualification_prep.R
\name{updateETAinNONMEMRecord}
\alias{updateETAinNONMEMRecord}
\title{Update ETA's in NONMEM record.}
\usage{
updateETAinNONMEMRecord(pharmpyModel, recordType, params)
}
\arguments{
\item{pharmpyModel}{pharmpy model}

\item{recordType}{record type to adapt}

\item{params}{CAMPSIS parameters}
}
\description{
Update ETA's in NONMEM record.
}
