% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonmem_update.R
\name{updateETAinNONMEMRecord}
\alias{updateETAinNONMEMRecord}
\title{Update ETA's in NONMEM record.}
\usage{
updateETAinNONMEMRecord(ctl, recordType, pmxtran)
}
\arguments{
\item{ctl}{NONMEM control stream}

\item{pmxtran}{pmxtran}

\item{recortType}{record type to adapt}
}
\description{
Update ETA's in NONMEM record.
}
