% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonmem_qualification_prep.R
\name{updateInitsNONMEM}
\alias{updateInitsNONMEM}
\title{Execute NONMEM. PsN will be called automatically by R. 
Prepared control stream 'model.mod' is executed automatically and NONMEM results
are returned in the form of a data frame.}
\usage{
updateInitsNONMEM(ctl)
}
\arguments{
\item{ctl}{path to control stream}
}
\value{
output path to PsN output
}
\description{
Execute NONMEM. PsN will be called automatically by R. 
Prepared control stream 'model.mod' is executed automatically and NONMEM results
are returned in the form of a data frame.
}
